+++ lib/src/core/parsers/models_parser.dart
@@ -36,7 +36,9 @@ class ModelsParser {
     final folders = folder.nodes.whereType<Folder>().toList();
     final items = folder.nodes.whereType<Item>().toList();
 
-    final foldersEntities = folders.expand((e) => parseFolder([...path, folder.name], e)).toList();
+    final foldersEntities = folders
+        .expand((e) => parseFolder([...path, folder.name.asName], e))
+        .toList();
 
     final List<Map<String, Value>> values = [];
 
@@ -58,7 +60,7 @@ class ModelsParser {
     final fieldsEntities = values.map(
       (e) {
         if (wrapWithClass) {
-          final name = _toName(e.values.first);
+          final name = _toName(e.values.first, identical(folder, root) ? folder.name.asName : '');
 
           return Field(
             name: name,
@@ -93,12 +95,12 @@ class ModelsParser {
     ];
   }
 
-  String _toName(Value value) {
+  String _toName(Value value, [String optionalPrefix = '']) {
     if (value is DoubleValue) {
       if (value.value.isInt) {
-        return value.value.toStringAsFixed(0);
+        return optionalPrefix + value.value.toStringAsFixed(0);
       } else {
-        return value.value.toString().snakeCase;
+        return optionalPrefix + value.value.toString().snakeCase;
       }
     } else {
       return value.toString().asName;