+++ lib/src/core/writers/build_context_extension_writer.dart
@@ -1,5 +1,6 @@
 import 'package:genuis/src/config/config.dart';
 import 'package:genuis/src/core/models/module.dart';
+import 'package:genuis/src/utils/string_extension.dart';
 
 class BuildContextExtensionWriter {
   BuildContextExtensionWriter();
@@ -7,11 +8,13 @@ class BuildContextExtensionWriter {
   void write(StringBuffer buffer, List<Module> modules) {
     buffer.writeln('extension ${Config.it.className}BuildContextExtension on BuildContext {');
     buffer.writeln(
-        '${Config.it.className} get ${Config.it.fieldName} => ${Config.it.className}.of(this);');
+      '${Config.it.className} get ${Config.it.fieldName} => ${Config.it.className}.of(this);',
+    );
 
     for (final module in modules) {
       buffer.writeln(
-          '${module.rootClass.classType} get ${module.config.name} => ${Config.it.fieldName}.${module.config.name};');
+        '${module.rootClass.classType} get ${module.config.name.asName} => ${Config.it.fieldName}.${module.config.name.asName};',
+      );
     }
 
     buffer.writeln('}');