+++ lib/src/core/writers/module/color_extension_writer.dart
@@ -45,12 +45,14 @@ class ColorExtensionWriter {
     }
     buffer.writeln(';');
 
-    buffer.writeln('$colorClassName(UI ui, $valueType $colorFieldName) : this._(');
-    for (final (name, field) in fields.iterable) {
-      buffer.writeln('$name: ui.${field.path.join('.')},');
+    for (final theme in module.rootClass.themes) {
+      buffer.writeln('$colorClassName.$theme($valueType $colorFieldName) : this._(');
+      for (final (name, field) in fields.iterable) {
+        buffer.writeln('$name: ${field.value(theme)},');
+      }
+      buffer.writeln('value: $colorFieldName,');
+      buffer.writeln(');');
     }
-    buffer.writeln('value: $colorFieldName,');
-    buffer.writeln(');');
 
     buffer.writeln('');
     buffer.writeln('$colorClassName lerp($colorClassName other, double t){');
@@ -63,12 +65,30 @@ class ColorExtensionWriter {
     buffer.writeln(');');
     buffer.writeln('}');
 
+    buffer.writeln('@override');
+    buffer.writeln('bool operator ==(Object other) {');
+    buffer.writeln('if (identical(this, other)) {');
+    buffer.writeln('return true;');
+    buffer.writeln('}');
+    buffer.writeln('if (other.runtimeType != runtimeType) {');
+    buffer.writeln('return false;');
+    buffer.writeln('}');
+    buffer.writeln('return other is $colorClassName');
+    buffer.writeln('&& other.$colorFieldName == $colorFieldName');
+    for (final name in fields.keys) {
+      buffer.writeln('&& other._$name == _$name');
+    }
+    buffer.writeln(';');
+    buffer.writeln('}');
+
     for (final name in fields.keys) {
       buffer.writeln(
-          '$recordClassName get $name => ${colorRecordClassName ?? ''}($colorFieldName, _$name);');
+        '$recordClassName get $name => ${colorRecordClassName ?? ''}($colorFieldName, _$name);',
+      );
     }
     buffer.writeln(
-        '$recordClassName colored(Color color) => ${colorRecordClassName ?? ''}($colorFieldName, color);');
+      '$recordClassName colored(Color color) => ${colorRecordClassName ?? ''}($colorFieldName, color);',
+    );
     buffer.writeln('}');
 
     if (colorRecordClassName != null) {
@@ -80,7 +100,7 @@ class ColorExtensionWriter {
     }
   }
 
-  void writeTextStyleExtensionClass(StringBuffer buffer, Map<String, Field> fields) {
+  void writeTextStyleExtensionClass(StringBuffer buffer, Module module, Map<String, Field> fields) {
     buffer.writeln('class $colorClassName extends TextStyle {');
     for (final name in fields.keys) {
       buffer.writeln('final Color _$name;');
@@ -122,15 +142,14 @@ class ColorExtensionWriter {
     buffer.writeln('overflow: style.overflow,');
     buffer.writeln(');');
 
-    buffer.writeln('$colorClassName(');
-    buffer.writeln('UI ui,');
-    buffer.writeln('TextStyle style');
-    buffer.writeln(') : this._(');
-    for (final (name, field) in fields.iterable) {
-      buffer.writeln('$name: ui.${field.path.join('.')},');
+    for (final theme in module.rootClass.themes) {
+      buffer.writeln('$colorClassName.$theme(TextStyle style) : this._(');
+      for (final (name, field) in fields.iterable) {
+        buffer.writeln('$name: ${field.value(theme)},');
+      }
+      buffer.writeln('style: style,');
+      buffer.writeln(');');
     }
-    buffer.writeln('style: style,');
-    buffer.writeln(');');
 
     buffer.writeln('');
     buffer.writeln('$colorClassName lerp($colorClassName other, double t){');
@@ -143,12 +162,28 @@ class ColorExtensionWriter {
     buffer.writeln(');');
     buffer.writeln('}');
 
+    buffer.writeln('@override');
+    buffer.writeln('bool operator ==(Object other) {');
+    buffer.writeln('if (identical(this, other)) {');
+    buffer.writeln('return true;');
+    buffer.writeln('}');
+    buffer.writeln('if (other.runtimeType != runtimeType) {');
+    buffer.writeln('return false;');
+    buffer.writeln('}');
+    buffer.writeln('return other is $colorClassName');
+    buffer.writeln('&& super == other');
+    for (final name in fields.keys) {
+      buffer.writeln('&& other._$name == _$name');
+    }
+    buffer.writeln(';');
+    buffer.writeln('}');
+
     buffer.writeln();
     for (final name in fields.keys) {
       buffer.writeln('TextStyle get $name => copyWith(color: _$name);');
     }
     buffer.writeln();
-    buffer.writeln('TextStyle colored(Color color) => copyWith(color: color);');
+    buffer.writeln('TextStyle colored(Color? color) => copyWith(color: color);');
 
     buffer.writeln('}');
   }