+++ lib/src/utils/string_extension.dart
@@ -13,7 +13,7 @@ extension StringExtension on String {
   String get withoutExtension => contains('.') ? substring(0, lastIndexOf('.')) : this;
   String get withoutFile => contains('/') ? substring(0, lastIndexOf('/')) : this;
 
-  String get forwardSlash => replaceAll('\\', '/');
+  String get forwardSlash => replaceAll(r'\', '/');
 
   String get asFolderPath {
     var result = forwardSlash;
@@ -29,18 +29,15 @@ extension StringExtension on String {
   }
 
   // Naming
-  String get camelCase => replaceAll(RegExp('[^A-Za-z0-9]'), ' ')
-      .split(' ')
-      .where((e) => e.isNotEmpty)
-      .map((e) => e.upperFirst)
-      .join()
-      .lowerFirst;
+  String get camelCase => replaceAll(
+    RegExp('[^A-Za-z0-9]'),
+    ' ',
+  ).split(' ').where((e) => e.isNotEmpty).map((e) => e.upperFirst).join().lowerFirst;
 
-  String get snakeCase => replaceAll(RegExp('[^A-Za-z0-9]'), ' ')
-      .split(' ')
-      .map((e) => e.toLowerCase())
-      .where((e) => e.isNotEmpty)
-      .join('_');
+  String get snakeCase => replaceAll(
+    RegExp('[^A-Za-z0-9]'),
+    ' ',
+  ).split(' ').map((e) => e.toLowerCase()).where((e) => e.isNotEmpty).join('_');
 
   String get exceptKeywords => _dartKeywords.contains(this) ? '${this}_' : this;