+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## 0.0.3
+
+* New: Add supported platforms to `pubspec.yaml`
+* New: Add topics to `pubspec.yaml`
+* Upd: Clean the `output_path` folder from old `.ui.dart` files before build
+* Fix: Module name in generated classes and getters
+* Fix: Recursive initialization in modules with `color: true`
+
 ## 0.0.2
 
-* Fix: issue with collecting themes
-* Fix: ignoring separators in json-keys
+* Fix: Issue with collecting themes
+* Fix: Ignoring separators in json-keys
 
 ## 0.0.1